Created new folder: Data. This folder contains our AppDbContext Class
This class  is going to be our database session/gateway. Whenever the app wants to query, insert, update, or delete data, it does it through this class

It contains a constructor with has 1 parameter: DbContextOptions<AppDbContext> options

This parameter is basically a config package that tells our EF Core how to connect to the database. The database connection settings (options) are usually configured in a Program.cs, so I will do that soon